Move cmk.gui.wato.mode -> cmk.gui.watolib
Fix import cycles introduced in 99a8862750f3.

  🐈
  ⌨️
