+++ cmk/gui/wato/pages/backup.py
@@ -15,8 +15,8 @@ from cmk.gui.i18n import _
 from cmk.gui.logged_in import user
 from cmk.gui.pages import AjaxPage, PageRegistry, PageResult
 from cmk.gui.type_defs import PermissionName
-from cmk.gui.wato.mode import ModeRegistry, WatoMode
 from cmk.gui.watolib.audit_log import log_audit
+from cmk.gui.watolib.mode import ModeRegistry, WatoMode
 
 
 def register(page_registry: PageRegistry, mode_registry: ModeRegistry) -> None:
