+++ cmk/gui/wato/pages/bulk_edit.py
@@ -24,7 +24,6 @@ from cmk.gui.plugins.wato.utils import (
 from cmk.gui.type_defs import ActionResult, PermissionName
 from cmk.gui.utils.flashed_messages import flash
 from cmk.gui.utils.transaction_manager import transactions
-from cmk.gui.wato.mode import ModeRegistry, redirect, WatoMode
 from cmk.gui.wato.pages.folders import ModeFolder
 from cmk.gui.watolib.host_attributes import (
     ABCHostAttribute,
@@ -37,6 +36,7 @@ from cmk.gui.watolib.hosts_and_folders import (
     Host,
     SearchFolder,
 )
+from cmk.gui.watolib.mode import ModeRegistry, redirect, WatoMode
 
 
 def register(mode_registry: ModeRegistry) -> None:
