+++ cmk/gui/wato/pages/bulk_import.py
@@ -47,16 +47,16 @@ from cmk.gui.valuespec import (
     TextInput,
     UploadOrPasteTextFile,
 )
-from cmk.gui.wato.mode import mode_url, redirect, WatoMode
 from cmk.gui.wato.pages.custom_attributes import ModeCustomHostAttrs
 from cmk.gui.wato.pages.folders import ModeFolder
 from cmk.gui.watolib.host_attributes import host_attribute_registry
 from cmk.gui.watolib.hosts_and_folders import folder_from_request
+from cmk.gui.watolib.mode import mode_url, redirect, WatoMode
 
 # Was not able to get determine the type of csv._reader / _csv.reader
 CSVReader = Any
 
-from cmk.gui.wato.mode import ModeRegistry
+from cmk.gui.watolib.mode import ModeRegistry
 
 
 def register(mode_registry: ModeRegistry) -> None:
