+++ cmk/gui/wato/pages/icons.py
@@ -25,8 +25,8 @@ from cmk.gui.type_defs import ActionResult, PermissionName
 from cmk.gui.utils.theme import theme
 from cmk.gui.utils.transaction_manager import transactions
 from cmk.gui.valuespec import Dictionary, DropdownChoice, IconSelector, ImageUpload
-from cmk.gui.wato.mode import ModeRegistry, redirect, WatoMode
 from cmk.gui.watolib.hosts_and_folders import make_action_link
+from cmk.gui.watolib.mode import ModeRegistry, redirect, WatoMode
 
 
 def register(mode_registry: ModeRegistry) -> None:
