+++ cmk/gui/wato/pages/object_parameters.py
@@ -28,7 +28,6 @@ from cmk.gui.plugins.wato.utils.context_buttons import make_service_status_link
 from cmk.gui.type_defs import PermissionName
 from cmk.gui.utils.html import HTML
 from cmk.gui.valuespec import Tuple, ValueSpecText
-from cmk.gui.wato.mode import ModeRegistry, WatoMode
 from cmk.gui.wato.pages.hosts import ModeEditHost, page_menu_host_entries
 from cmk.gui.watolib.check_mk_automations import analyse_host, analyse_service
 from cmk.gui.watolib.hosts_and_folders import (
@@ -37,6 +36,7 @@ from cmk.gui.watolib.hosts_and_folders import (
     folder_preserving_link,
     Host,
 )
+from cmk.gui.watolib.mode import ModeRegistry, WatoMode
 from cmk.gui.watolib.rulesets import AllRulesets, Rule, Ruleset
 from cmk.gui.watolib.rulespecs import (
     get_rulegroup,
