+++ cmk/gui/watolib/mode/_registry.py
