+++ tests/unit/cmk/gui/wato/test_mode.py
@@ -10,8 +10,8 @@ import pytest
 from cmk.gui.breadcrumb import BreadcrumbItem
 from cmk.gui.plugins.wato.utils.main_menu import MainModuleTopicHosts
 from cmk.gui.type_defs import Icon, PermissionName
-from cmk.gui.wato.mode import _base, WatoMode
 from cmk.gui.watolib.main_menu import ABCMainModule, MainModuleRegistry, MainModuleTopic
+from cmk.gui.watolib.mode import _base, WatoMode
 
 module_registry = MainModuleRegistry()
